refactor: move checkers page to app
